Automatically provision io2 EBS volumes instead of io1
They have 100x Higher Durability and 10x More IOPS/GiB at the same cost.

My only concern so far is related to regional coverage, since they're not available in all AWS regions yet.
We need to find a way to only do this on the supported regions, if possible without hardcoding the list of supported regions.